python/django-rest-framework/pyproject.toml
[project]
name = "django-rest-framework"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "djangorestframework>=3.15",
]
